% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ged_fit}
\alias{ged_fit}
\title{Fit Generalized Error Distribution to a vector of returns/stock prices.}
\usage{
ged_fit(vec)
}
\arguments{
\item{vec}{A numeric vector of data.}
}
\value{
A list with the following elements:
\describe{
\item{params}{A numeric vector of length 3 containing the fitted GED parameters: shape, scale, and location.}
\item{aic}{The Akaike Information Criterion (AIC) for the fitted model.}
\item{bic}{The Bayesian Information Criterion (BIC) for the fitted model.}
}
}
\description{
This function fits the Generalized Error Distribution (GED) to a given data vector using the \code{ged_fit} function from
the \code{fGarch} package. It returns the estimated parameters along with the AIC and BIC values for the fitted
distribution.
}
\examples{
stock_prices <- c(10, 11, 12, 13, 14)
returns <- diff(log(stock_prices))
ged_fit(returns)

}
