% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{skew.t_fit}
\alias{skew.t_fit}
\title{Fit Skewed Student-t Distribution to a vector of returns/stock prices.}
\usage{
skew.t_fit(vec)
}
\arguments{
\item{vec}{A numeric vector of data.}
}
\value{
A list with the following elements:
\describe{
\item{params}{A numeric vector of length 4 containing the fitted Skewed Student-t parameters: degrees of freedom, skewness, scale, and location.}
\item{aic}{The Akaike Information Criterion (AIC) for the fitted model.}
\item{bic}{The Bayesian Information Criterion (BIC) for the fitted model.}
}
}
\description{
This function fits the Skewed Student-t Distribution to a given data vector using the \code{skew.t_fit} function from
the \code{fGarch} package. It returns the estimated parameters along with the AIC and BIC values for the fitted
distribution.
}
\examples{
stock_prices <- c(10, 11, 12, 13, 14)
returns <- diff(log(stock_prices))
skew.t_fit(returns)

}
