% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sym.hd_fit}
\alias{sym.hd_fit}
\title{Fit a Symmetric Hyperbolic Distribution to a vector of return/stock prices.}
\usage{
sym.hd_fit(vec)
}
\arguments{
\item{vec}{a numeric vector containing the symmetric data to be fitted.}
}
\value{
a list containing the following elements:
\describe{
\item{par}{a numeric vector of length 4 containing the estimated values for the parameters of the
fitted distribution: alpha (scale), mu (degrees of freedom), sigma (standard deviation), and gamma (skewness).}
\item{aic}{the Akaike information criterion (AIC) value for the fitted distribution.}
\item{bic}{the Bayesian information criterion (BIC) value for the fitted distribution.}
}
}
\description{
This function fits a Symmetric Hyperbolic distribution to a  data vector using the
\code{fit.hypuv} function from the \code{ghyp} package. It returns the estimated parameters along with
the AIC and BIC values for the fitted distribution.
}
\examples{
\dontrun{
stock_prices <- c(10, 11, 12, 13, 14)
returns <- diff(log(stock_prices))
sym.hd_fit(returns)
}

}
\seealso{
\code{\link{norm_fit}}, \code{\link{t_fit}}, \code{\link{cauchy_fit}}, \code{\link{ghd_fit}}, \code{\link{hd_fit}},
\code{\link{sym.ghd_fit}}, \code{\link{vg_fit}}, \code{\link{sym.vg_fit}}, \code{\link{nig_fit}},
\code{\link{ged_fit}}, \code{\link{skew.t_fit}}, \code{\link{skew.normal_fit}}, \code{\link{skew.ged_fit}}
}
