% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{annual_return}
\alias{annual_return}
\title{Compute Annual Returns of a Vector.}
\usage{
annual_return(vec)
}
\arguments{
\item{vec}{a numeric vector of asset returns as an xts object with dates as rownames.}
}
\value{
A numeric vector of annual returns.
}
\description{
This function takes a vector of asset returns and computes annual returns.
}
\examples{
\dontrun{
# Compute annual returns of an asset vector
asset_returns_xts <- xts(c(0.05, -0.03, 0.02, -0.01, 0.04, -0.02, 0.01),
order.by = as.Date(c("2023-05-01", "2023-05-02", "2023-05-03",
"2023-05-04", "2023-05-05", "2023-05-06",
"2023-05-07")))
annual_return(asset_returns_xts)
}

}
\seealso{
\code{\link{weekly_return}}, \code{\link{monthly_return}}
}
