% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{weekly_return}
\alias{weekly_return}
\title{Compute Weekly Returns of a Vector.}
\usage{
weekly_return(vec)
}
\arguments{
\item{vec}{a numeric vector of asset returns.}
}
\value{
A numeric vector of weekly returns.
}
\description{
This function takes a numeric vector of asset returns and computes weekly returns.
}
\note{
The input data must be an xts object with dates as rownames.
}
\examples{
\dontrun{
# Compute weekly returns of an asset vector
require(xts)
asset_returns_xts <- xts(c(0.05, -0.03, 0.02, -0.01, 0.04, -0.02, 0.01),
                         order.by = as.Date(c("2023-05-01", "2023-05-02", "2023-05-03",
                                             "2023-05-04", "2023-05-05", "2023-05-06",
                                             "2023-05-07")))
weekly_return(asset_returns_xts)
}

}
\seealso{
\code{\link{monthly_return}}, \code{\link{annual_return}}
}
